requirements/base.txt
asgiref==3.3.4
backports.zoneinfo==0.2.1
Django==3.2.12
django-cleanup==6.0.0
mysql-connector-python==8.0.23
mysqlclient==2.0.3
Pillow==9.0.1
protobuf==3.15.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.2
